requirements.txt
# Python Backend for CIM Automation Platform
# Using versions with pre-built Windows wheels (no C compiler needed)

# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25

# Validation & Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# File Handling
python-multipart>=0.0.6

# Data Processing (use >= to get pre-built wheels)
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# AI Integration
google-generativeai>=0.3.0

# PDF Generation (pure Python)
fpdf2>=2.7.0
jinja2>=3.1.0

# Streamlit Frontend
streamlit>=1.30.0
streamlit-extras>=0.3.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
